✨ feat(client): jsx 형식을 tsx로 변환
