--- packages/client/src/App.js
@@ -1,15 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-import NavigationBar from './component/AppBar/AppBar/NavigationBar.tsx';
-
-function App() {
-  return (
-    <body bgcolor='#50bcdf'>
-      <div className="App">
-        <NavigationBar />
-      </div>
-    </body>
-  );
-}
-
-export default App;
