+++ packages/client/src/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import logo from './logo.svg';
+import './App.css';
+import NavigationBar from './components/bar/NavigationBar';
+
+function App() {
+  return (
+    <div className="App">
+      <NavigationBar />
+    </div>
+  );
+}
+
+export default App;
